Get the latest test and conver to junit format
