+++ java/pom.xml
@@ -1,4 +1,5 @@
-<project xmlns="https://maven.apache.org/POM/4.0.0" xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="https://maven.apache.org/POM/4.0.0" xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="https://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.google</groupId>
     <artifactId>diff-match-patch</artifactId>
@@ -80,9 +81,13 @@
 
             <plugin>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.20</version>
+                <version>2.21.0</version>
                 <configuration>
                     <surefire.useFile>false</surefire.useFile>
+                    <includes>
+                        <include>**/diff_match_patch_test.java</include>
+                        <include>**/*Test</include>
+                    </includes>
                 </configuration>
             </plugin>
         </plugins>
